refactor(lupaTool): extrair CameraSVG e sincronizar zoom entre canvas e overlay.
